% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read file with FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string with the name of the file to read, see
notes.}
}
\value{
A data frame with data readed from the csv file, or an error if the
  file does not exists.
}
\description{
\code{fars_read()} function reads data from a \code{*.csv} file, stored on disk,
from the \strong{US National Highway Traffic Safety Administration's}
\emph{Fatality Analysis Reporting System} (FARS), which is a nationwide census,
providing the American public yearly data, regarding fatal injuries suffered in
motor vehicle traffic crashes. If the filename does not exist the function will
stop and an error will be thrown.
}
\details{
For more information, see:
\itemize{
  \item{\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}}
  \item{\url{https://en.wikipedia.org/wiki/Fatality_Analysis_Reporting_System}}
}
}
\note{
To generate file name use: \code{\link{make_filename}}
}
\examples{
\dontrun{fars_read("accident_2014.csv.bz2")}
library(readr)
\dontrun{fars_read("data/accident_2014.csv.bz2")}

}
\seealso{
\link{make_filename}
}
