% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{farsyrsum}
\alias{farsyrsum}
\title{Summarized yearly FARS data for 2013 to 2015}
\format{
A data frame with columns:
\describe{
 \item{MONTH}{The month id for the year, 1 = January, etc.}
 \item{2013}{The number of fatality incidents for year 2013.}
 \item{2014}{The number of fatality incidents for year 2014.}
 \item{2015}{The number of fatality incidents for year 2015.}
}
}
\source{
US National Highway Traffic Safety Administration's Fatality Analysis
 Reporting System (FARS). \url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
}
\usage{
data(farsyrsum)
}
\description{
Summarized yearly FARS data for 2013 to 2015
}
\examples{
{
 data(farsyrsum)
}
}
\keyword{datasets}
